chore: move marker linkage files to structed path
