+++ genshin-map-data/genshin-map-data-core/genshin-map-data-core-mapper/src/main/resources/mapper/MBPMarkerLinkageMapper.xml
@@ -16,7 +16,7 @@
         <result column="to_id" property="toId"/>
         <result column="link_action" property="linkAction"/>
         <result column="link_reverse" property="linkReverse"/>
-        <result column="path" property="path" typeHandler="site.yuanshen.data.helper.MarkerLinkagePathTypeHandler"/>
+        <result column="path" property="path" typeHandler="site.yuanshen.data.helper.marker.linkage.MarkerLinkagePathTypeHandler"/>
         <result column="extra" property="extra" typeHandler="site.yuanshen.handler.MBPJsonObjectTypeHandler"/>
     </resultMap>
